% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{addElement}
\alias{addElement}
\title{Add an element to a given list}
\usage{
addElement(givenList, elementToAdd)
}
\arguments{
\item{givenList}{Given list}

\item{elementToAdd}{Element to add}
}
\value{
New list with added element
}
\description{
Add an element to a given list
}
