% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_plots.R
\name{align_plots}
\alias{align_plots}
\title{Align plots in one column}
\usage{
align_plots(p, save = FALSE, fname = NULL)
}
\arguments{
\item{p}{A \code{list} of plots}

\item{save}{Should the outcome be saved to a \code{.png}-file?}

\item{fname}{Provide a file name for the save file. If none is provided, one
is produced from the plot list name.}
}
\description{
Arranges plots from \code{ggplot} into one column. The use case is to have
several plots which need to be arranged along a time line. With other approaches,
labels and ticks on the vertical axes are taken into account, because they are
part of the plot. With the \code{grid} package, it is possible to align them
along the x-axis.
}
