% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angles.R
\name{angles}
\alias{angles}
\alias{r2d}
\alias{d2r}
\title{Angle transformations}
\usage{
r2d(x)

d2r(x)
}
\arguments{
\item{x}{An angle}
}
\value{
The transformed angle
}
\description{
Convenience functions to transform angles from \code{radian} to \code{degree}, and from \code{degree} to \code{radian}.
}
\examples{

# Radian to degree
r2d(2 * pi)

# Degree to radian
d2r(360)

}
