% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_datetime.R
\name{as_datetime}
\alias{as_datetime}
\alias{uxt}
\alias{win2uxt}
\title{Convert an object to a date or date-time with defaults}
\usage{
uxt(time, origin = "1970-01-01", tz = "Europe/Berlin")

win2uxt(time, origin = "1601-01-01", tz = "Europe/Berlin")
}
\arguments{
\item{time}{A numerical value.}

\item{origin}{a Date object, or something which can be coerced by as.Date(origin, ...) to such an object (default: the Unix epoch of "1970-01-01" for \code{uxt}, and "1601-01-01" for \code{w2u}).}

\item{tz}{a time zone name (default: Europe/Berlin)}
}
\value{
a vector of Date objects corresponding to x.
}
\description{
Wrapper for \code{lubridate::as_datetime} with defaults for \code{unixtime} and \code{windowstime}.
}
\details{
\emph{Unix Time} := seconds since 1970-01-01 at 00:00:00.0
\emph{Windows Time} := 100 ns since 1601-01-01 at 00:00:00.0

Note:   difftime("1970-01-01", "1601-01-01", units="secs") is 692 s than what would be expected.
Reason: The value below is the diff in UTC (extended back to 1601), while
        difftime() considers the time jump of 6:32 min on 1893-04-01 in
        Berlin when switching from local time to CET.
}
\examples{

# convert unixtime to a date object
uxt(1550242175.5)

}
