% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{buildPeriodVector}
\alias{buildPeriodVector}
\title{Build vector filled with repeating 1, 2, 3...}
\usage{
buildPeriodVector(periodLength, numOfPeriods)
}
\arguments{
\item{periodLength}{How many times to repeat each number}

\item{numOfPeriods}{Number of periods}
}
\value{
Vector filled with repeating 1, 2, 3...
}
\description{
Build vector filled with repeating 1, 2, 3...
}
