% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{calcLogChanges}
\alias{calcLogChanges}
\title{Calculate vector of log percent changes of input vector}
\usage{
calcLogChanges(v)
}
\arguments{
\item{v}{Given vector}
}
\value{
Vector of log percent changes
}
\description{
Calculate vector of log percent changes of input vector
}
