% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{calcPercentChanges}
\alias{calcPercentChanges}
\title{Calculate vector of simple percent changes of input vector}
\usage{
calcPercentChanges(v)
}
\arguments{
\item{v}{Given vector}
}
\value{
Vector of simple percent changes
}
\description{
Calculate vector of simple percent changes of input vector
}
