% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{calcPercentInsideInclusive}
\alias{calcPercentInsideInclusive}
\title{Calculate percentage of vector's elements inside given range inclusive}
\usage{
calcPercentInsideInclusive(v, range)
}
\arguments{
\item{v}{given vector}

\item{range}{given range}
}
\value{
percentage of vector's elements inside range
}
\description{
Calculate percentage of vector's elements inside given range inclusive
}
