% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{chartHist}
\alias{chartHist}
\title{Chart nice histogram}
\usage{
chartHist(v, numOfBars = 50, color = rgb(0, 0, 1, 1/2),
  xlab = deparse(substitute(v)), main = "Frequency")
}
\arguments{
\item{v}{Data vector}

\item{numOfBars}{Number of bars}

\item{color}{Color of histogram}

\item{xlab}{Label of x}

\item{main}{Title}
}
\value{
Histogram
}
\description{
Chart nice histogram
}
