% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.name.R
\name{extract.name}
\alias{extract.name}
\title{Extraction of a name from path or file name}
\usage{
extract.name(fname)
}
\arguments{
\item{fname}{File name}
}
\value{
A string with the file name without extension.
}
\description{
Extract a name from a path or file name excluding file extension. Match anything up to but not including last dot or whole string without any dots.
}
\examples{

extract_name("D:\\\\Home\\\\projects\\\\awar.d\\\\r-workspace\\\\functions.R")
}
