% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{getIndicesOfMax}
\alias{getIndicesOfMax}
\title{Get indices of max element of matrix}
\usage{
getIndicesOfMax(m)
}
\arguments{
\item{m}{given matrix}
}
\value{
indices of max element of matrix
}
\description{
Get indices of max element of matrix
}
