% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{getLengths}
\alias{getLengths}
\title{Return vector of lengths of list elements}
\usage{
getLengths(l, atoms = c("atom"))
}
\arguments{
\item{l}{List}

\item{atoms}{Vector of class names that lengths are assumed to be 1}
}
\value{
Vector of lengths
}
\description{
Return vector of lengths of list elements
}
