% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{getSurfColors}
\alias{getSurfColors}
\title{Return set of colors for 3d surface
Author: ucgamdo@ucl.ac.uk}
\usage{
getSurfColors(z, col = cm.colors(40))
}
\arguments{
\item{z}{matrix of z-values of the surface}

\item{col}{palette}
}
\value{
vector of colors for 3d surface
}
\description{
Return set of colors for 3d surface
Author: ucgamdo@ucl.ac.uk
}
