% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{KKKV}
\alias{KKKV}
\title{Make a function of four arguments take a vector as a fourth argument}
\usage{
KKKV(f, k1, k2, k3, v)
}
\arguments{
\item{f}{Function}

\item{k}{First argument as a constant}

\item{v1}{Second argument as a constant}

\item{v2}{Third argument as a constant}

\item{v3}{Fourth argument as a vector}
}
\value{
Vector of outputs
}
\description{
Make a function of four arguments take a vector as a fourth argument
}
