% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{KV}
\alias{KV}
\title{Make a function of two arguments take a vector as a second argument}
\usage{
KV(f, k, v)
}
\arguments{
\item{f}{Function}

\item{k}{First argument as a constant}

\item{v}{Second argument as a vector}
}
\value{
Vector of outputs
}
\description{
Make a function of two arguments take a vector as a second argument
}
