% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{KVV}
\alias{KVV}
\title{Make a function of three arguments take vectors as a second and third arguments}
\usage{
KVV(f, k, v1, v2)
}
\arguments{
\item{f}{Function}

\item{k}{First argument as a constant}

\item{v1}{Second argument as a vector}

\item{v2}{Third argument as a vector}
}
\value{
Vector of outputs
}
\description{
Make a function of three arguments take vectors as a second and third arguments
}
