% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{KVVV}
\alias{KVVV}
\title{Make a function of four arguments take vectors as a second, third and fourth arguments}
\usage{
KVVV(f, k, v1, v2, v3)
}
\arguments{
\item{f}{Function}

\item{k}{First argument as a constant}

\item{v1}{Second argument as a vector}

\item{v2}{Third argument as a vector}

\item{v3}{Fourth argument as a vector}
}
\value{
Vector of outputs
}
\description{
Make a function of four arguments take vectors as a second, third and fourth arguments
}
