% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{matrixSwitch}
\alias{matrixSwitch}
\title{Switch that takes expression and results as vectors (helper.R)}
\usage{
matrixSwitch(expression, ...)
}
\arguments{
\item{expression}{Expression vector}

\item{...}{Cases and vectors of results. Results must be same class}
}
\value{
Result[i] of case that matches expression[i]
}
\description{
Switch that takes expression and results as vectors (helper.R)
}
\examples{
matrixSwitch(c("a","b","c"), a = 1:3, b = 11:13, c = 101:103, NA)

}
