% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmatch.R
\name{nmatch}
\alias{nmatch}
\alias{\%nin\%}
\alias{\%nix\%}
\title{Negative match}
\usage{
nmatch(x, table)

x \%nin\% table

x \%nix\% table
}
\arguments{
\item{x}{A vector}

\item{table}{The data set to be queried.}
}
\value{
logical vector or index vector
}
\description{
While \code{match} returns the position of \code{y} where an element of \code{x} was found, \code{nmatch} and \code{\%nin\%} returns the position of \code{x} which were not found in \code{y}. It is not possible to give the position of \code{y} where something was \emph{not} found.
}
\details{
For \code{\%nix\%}, the index of the non-intersecting values for \code{x} is returned.
}
\examples{

x <- c(1, 2, 3)
y <- c(3, 4, 5)

nmatch(x, y)
x \%nix\% y

}
