% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norml.R
\name{norml}
\alias{norml}
\title{Normalize a vector between upper and lower bound.}
\usage{
norml(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{A vector}

\item{lower}{Lower bound: the smallest value in the vector will be this value.}

\item{upper}{Upper bound: the largest value in the vector will be this value.}
}
\value{
A vector
}
\description{
Given a vector, this function linearly transforms the vector's values between
an upper and a lower bound.
}
\details{
See the \href{http://stats.stackexchange.com/questions/178626/how-to-normalize-data-between-1-and-1}{algorithm} for details
}
\examples{
x <- rnorm(10)
norml(x)
norml(x, -8, 50)
}
