% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rledt.R
\name{rledt}
\alias{rledt}
\title{Enhanced \code{rle} for \code{data.table}s}
\usage{
rledt(x, na.collapse = TRUE)
}
\arguments{
\item{x}{A \code{data.table}}

\item{na.collapse}{Should \code{NA}s be indexed each, or treated as a block.}
}
\value{
A \code{data.table}
}
\description{
\code{rle} for \code{data.table}s with additional information.
}
\examples{

}
