% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSD.R
\name{rowSD}
\alias{rowSD}
\title{\code{standard deviation} for for rows}
\usage{
rowSD(x, na.rm = FALSE)
}
\arguments{
\item{x}{A \code{matrix} or coercible \code{list}.}

\item{na.rm}{Passed to \code{sd}}
}
\value{
A vector
}
\description{
Gives the \code{standard deviation} for the rows of a \code{matrix} or lists that can be coerced to a \code{matrix}.
}
\examples{

}
