% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sec2str.R
\name{sec2str}
\alias{sec2str}
\title{Converts a numerical value to the format \code{HH:MM::SS.MS}}
\usage{
sec2str(time, digits = 2)
}
\arguments{
\item{time}{A numerical value.}

\item{digits}{Number of decimal digits of the string.}
}
\value{
A string corresponding to x.
}
\description{
Converts seconds to a \emph{HH:MM:SS.MS} format. It handles
milliseconds well, is much short than the alternatives, and returns a string
instead of a date class.
}
\examples{

sec2str(3612.5)

}
