% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slct.R
\name{slct}
\alias{slct}
\title{Find column indices or names with \code{regex}}
\usage{
slct(pattern, names_df, names = T)
}
\arguments{
\item{pattern}{\code{regex}-pattern to match the column names.}

\item{names_df}{Names of the data.frame to be matched.}

\item{names}{Should the actual names be returned?}
}
\value{
Either the index of the matched column names, or their index.
}
\description{
Find column indices or names by matching them against a regular expression.
}
\examples{

x <- data.frame(a = 1, b = 2, ab = 3)
slct("a", names(x))

}
