% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart.round.R
\name{smart.round}
\alias{smart.round}
\title{Round double so that the sum stays the same.}
\usage{
smart.round(x, digits = 0)
}
\arguments{
\item{x}{A \code{vector}, \code{matrix}, or numerical \code{data.frame}.}

\item{digits}{\code{integer}: Round to the \code{nth}-decimal place}
}
\value{
same as input
}
\description{
Rounding numbers of the type \code{double} will change the sum of these
numbers. E.g. for a stochastic vector where the sum must be \code{1} this is
not acceptable. After calling \code{round}, \code{smart.round} changes the
data so that it preserves the sum of the input, with the minimal difference
to the input data.
}
\details{
\itemize{
  \item{\href{https://stackoverflow.com/questions/792460/how-to-round-floats-to-integers-while-preserving-their-sum}{algorithm}}
  \item{\href{https://stackoverflow.com/questions/32544646/round-vector-of-numerics-to-integer-while-preserving-their-sum}{R-implementation}}
  }
}
\examples{
set seed(42)
x <- rnorm(10)
y <- smart.round(x, 3)
round(sum(x), 3) == sum(y)
}
