% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{stretch}
\alias{stretch}
\title{Stretch elementary argument to reach length of others}
\usage{
stretch(a, ...)
}
\arguments{
\item{a}{Argument to stretch}

\item{...}{Other arguments}
}
\value{
Stretched argument with rep(argument, length of others)
}
\description{
Stretch elementary argument to reach length of others
}
\examples{
x <- 1
y <- c("a", "b")
z <- "c"
x <- stretch(x, y, z)
y <- stretch(y, x, z)
z <- stretch(z, y, x)

}
