% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{sumPeriod}
\alias{sumPeriod}
\title{Aggregate data of data frame summing up given periods}
\usage{
sumPeriod(data, periodLength)
}
\arguments{
\item{data}{Given data frame}

\item{periodLength}{Length of period}
}
\value{
Aggregated data frame
}
\description{
Aggregate data of data frame summing up given periods
}
