% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{trans_matrix}
\alias{trans_matrix}
\title{Generate a transition matrix}
\usage{
trans_matrix(x, conv = FALSE, cast.out = FALSE, lag = 1,
  prob = FALSE)
}
\arguments{
\item{x}{A vector, a matrix, or a data.frame with numerical values.}

\item{conv}{Converts the input to a matrix of \code{dim(nrow=1, ncol=length(x))}.}

\item{cast.out}{String. Casts the output to a \code{data.frame}, \code{data.table} or \code{matrix}.}

\item{lag}{Looks at the nth previous element in the markov chain vector.}

\item{prob}{If set \code{TRUE}, returns relative frequencies, otherwise they are absolute.}
}
\value{
A table object, or what in \code{cast.out} was defined.
}
\description{
Generates a transition matrix from the input.
}
\examples{

x <- sample(toupper(letters[1:5]), 100, replace = TRUE)
y <- trans_matrix(x, conv = TRUE, cast.out = "matrix", prob = TRUE)

}
