% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlpr.R
\name{vectorSwitch}
\alias{vectorSwitch}
\title{Common switch that takes expression as vector (helper.R)}
\usage{
vectorSwitch(expression, ...)
}
\arguments{
\item{expression}{Expression vector}

\item{...}{Cases and results. Results must be same class}
}
\value{
Result of case that matches expression[i]
}
\description{
Common switch that takes expression as vector (helper.R)
}
\examples{
vectorSwitch(c("a","c"), a = 1, b = 2, c = 3, 4)

}
