% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{Writes data to a file}
\usage{
w(x, fname = "test", sep = ";", dec = ".")
}
\arguments{
\item{x}{Data of type \code{list}}

\item{fname}{Name for the file. Default is "test". If not provided, \code{.csv}
will be appended.}

\item{sep}{CSV separator}

\item{dec}{Decimal separator}
}
\description{
The use case for this function is to quickly write out a data.frame into a
test file. It uses data.table's \code{fwrite}, which is faster than \code{write.table}
and handels large values (floats) much better.
}
\examples{
w("my_data")
w("my_data", "../analysis/my_data.csv")
}
