% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipper.R
\name{zipper}
\alias{zipper}
\title{Zipper}
\usage{
zipper(..., fun = c, vec = FALSE)
}
\arguments{
\item{...}{An arbitrary number of vectors}

\item{fun}{Function to be used to operate with the tupels.}

\item{vec}{If \code{TRUE}, a single vector is returned.}
}
\value{
Either a vector or a list of the zipped vector elements.
}
\description{
Inspired by Python's \code{zipper}. Taken two or more vectors, this function
arranges them like so: \code{(n11, n21), (n12, n22), (n13, n23)}. Functions
other than \code{c} can be used to work with the tupels.
}
\examples{
x <- 1:5
y <- 6:10
zip(x, y, vec = TRUE)
zip(x, y, fun = seq, vec = TRUE)
}
