% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{trans_matrix}
\alias{trans_matrix}
\title{Make a transition matrix.}
\usage{
trans_matrix(x, conv = FALSE, cast.out = FALSE, lag = 1, prob = FALSE)
}
\arguments{
\item{x}{a vector, a matrix, or a data.frame with numerical values}

\item{conv}{converts the input to a matrix of \code{dim(nrow=1, ncol=length(x))}}

\item{cast.out}{casts the output to a data.frame/data.table/matrix.}

\item{lag}{looks at the nth previous element in the markov chain vector}

\item{prob}{if set \code{TRUE}, returns relative frequencies, otherwise they are absolute}
}
\value{
a table object, or what in \code{cast.out} was defined
}
\description{
Make a transition matrix.
}
\examples{
x <- sample(toupper(letters[1:5]), 100, replace = T)
y <- trans_matrix(x, conv=T, cast.out="matrix", prob=T)
}
