% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w.R
\name{w}
\alias{w}
\title{Writing a csv with defaults}
\usage{
w(x = NULL, fname = NULL, sep = ";", dec = ",",
  overwrite = FALSE)
}
\arguments{
\item{x}{Any list of same length vectors, e.g. \code{data.frame} and \code{data.table}}

\item{fname}{Name for the file. Default is the name of the object to be
written.}

\item{sep}{CSV separator}

\item{dec}{decimal separator}

\item{overwrite}{Should existing files of the same name be overwritten?}
}
\description{
Write a data.frame to disk using \code{data.table::fwrite} with sensible
defaults. The default name of the file is the name of the object itself.
}
\examples{

df <- list(a = 1, b = 2)
w(df)

}
